+++ src/AvroConvert/AvroObjectServices/Schema/DateSchema.cs
@@ -42,7 +42,7 @@ namespace SolTechnology.Avro.AvroObjectServices.Schema
         }
 
 
-        internal override object ConvertToLogicalValue(object baseValue, LogicalTypeSchema schema, Type type)
+        internal override object ConvertToLogicalValue(object baseValue, LogicalTypeSchema schema, Type readType)
         {
             var noDays = (int)baseValue;
             return DateTimeExtensions.UnixEpochDateTime.AddDays(noDays);