+++ src/AvroConvert/AvroObjectServices/Schema/UuidSchema.cs
@@ -34,7 +34,7 @@ namespace SolTechnology.Avro.AvroObjectServices.Schema
         internal override AvroType Type => AvroType.Logical;
         internal override TypeSchema BaseTypeSchema { get; set; }
         internal override string LogicalTypeName => LogicalTypeEnum.Uuid;
-        internal override object ConvertToLogicalValue(object baseValue, LogicalTypeSchema schema, Type type)
+        internal override object ConvertToLogicalValue(object baseValue, LogicalTypeSchema schema, Type readType)
         {
             return Guid.Parse((string)baseValue);
         }